# docs/openapi.yaml
openapi: 3.0.0
info:
  title: PlagiaScan API
  description: API for PlagiaScan Plagiarism Detection System
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

paths:
  /auth/login:
    post:
      summary: User Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string

  /documents:
    post:
      summary: Upload a document
      tags: [Documents]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        201:
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
    get:
      summary: List user documents
      tags: [Documents]
      responses:
        200:
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'

  /documents/{id}:
    get:
      summary: Get document details
      tags: [Documents]
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
    delete:
      summary: Delete a document
      tags: [Documents]
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        204:
          description: Document deleted

  /scans:
    post:
      summary: Initiate a plagiarism scan
      tags: [Scans]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                document_id:
                  type: integer
                check_web:
                  type: boolean
                  default: false
                check_internal:
                  type: boolean
                  default: true
      responses:
        202:
          description: Scan initiated
          content:
            application/json:
              schema:
                type: object
                properties:
                  scan_id:
                    type: integer
                  status:
                    type: string

  /scans/{id}:
    get:
      summary: Get scan status and results
      tags: [Scans]
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        200:
          description: Scan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'

components:
  schemas:
    Document:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        status:
          type: string
          enum: [pending, processing, indexed, failed]
        created_at:
          type: string
          format: date-time

    ScanResult:
      type: object
      properties:
        id:
          type: integer
        document_id:
          type: integer
        status:
          type: string
          enum: [queued, scanning, completed, failed]
        score:
          type: number
          format: float
          description: Overall plagiarism score (0-100)
        matches:
          type: array
          items:
            type: object
            properties:
              source:
                type: string
              similarity:
                type: number
              fragments:
                type: array
                items:
                  type: object
                  properties:
                    text:
                      type: string
                    start_index:
                      type: integer
                    end_index:
                      type: integer